Collapse FileInputContainer into App
